+++ examples/resources/rubrik_aws_cnp_account/import-by-string-id.tf
@@ -1,4 +1,4 @@
 import {
   to = rubrik_aws_cnp_account.account
-  id = "3553bc74-7061-40e3-bac5-d2639e58bb7e-external-id"
+  id = "3553bc74-7061-40e3-bac5-d2639e58bb7e:external-id"
 }