+++ internal/provider/framework_checkdestroys_test.go
@@ -45,6 +45,38 @@ func awsAccountCheckDestroy(ctx context.Context) func(*terraform.State) error {
 	}
 }
 
+// awsCnpAccountCheckDestroy verifies that all aws_cnp_account resources have
+// been deleted.
+func awsCnpAccountCheckDestroy(ctx context.Context) func(*terraform.State) error {
+	return func(s *terraform.State) error {
+		client, err := testClient(ctx)
+		if err != nil {
+			return err
+		}
+
+		for _, rs := range s.RootModule().Resources {
+			if rs.Type != "polaris_aws_cnp_account" && rs.Type != "rubrik_aws_cnp_account" {
+				continue
+			}
+
+			id, err := uuid.Parse(rs.Primary.ID)
+			if err != nil {
+				return err
+			}
+
+			_, err = aws.Wrap(client).AccountByID(ctx, id)
+			if err == nil {
+				return fmt.Errorf("aws_cnp_account %s still exists", id)
+			}
+			if !errors.Is(err, graphql.ErrNotFound) {
+				return err
+			}
+		}
+
+		return nil
+	}
+}
+
 // customRoleCheckDestroy verifies that all custom_role resources have been
 // deleted.
 func customRoleCheckDestroy(ctx context.Context) func(*terraform.State) error {