+++ internal/provider/framework_provider.go
@@ -123,6 +123,8 @@ func (p *FrameworkProvider) Resources(ctx context.Context) []func() resource.Res
 	tflog.Trace(ctx, "FrameworkProvider.Resources")
 
 	return []func() resource.Resource{
+		newAwsCnpAccountResource,
+		newPolarisAwsCnpAccountResource,
 		newCustomRoleResource,
 		newPolarisCustomRoleResource,
 		newRoleAssignmentResource,